# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Theory of Change API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /theories/{theoryId}/documents:
    get:
      operationId: listTheoryDocuments
      tags: [theories]
      summary: List documents for a theory
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TheoryDocument"
    post:
      operationId: createTheoryDocument
      tags: [theories]
      summary: Save document metadata after upload
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTheoryDocument"
      responses:
        "201":
          description: Document saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TheoryDocument"
  /theories/{theoryId}/documents/{docId}:
    delete:
      operationId: deleteTheoryDocument
      tags: [theories]
      summary: Delete a document
      parameters:
        - name: theoryId
          in: path
          required: true
          schema:
            type: integer
        - name: docId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    TheoryDocument:
      type: object
      properties:
        id:
          type: integer
        theoryId:
          type: integer
        name:
          type: string
        objectPath:
          type: string
        contentType:
          type: string
        size:
          type: integer
        uploadedAt:
          type: string
          format: date-time
      required:
        - id
        - theoryId
        - name
        - objectPath
        - contentType
        - size
        - uploadedAt
    CreateTheoryDocument:
      type: object
      properties:
        name:
          type: string
        objectPath:
          type: string
        contentType:
          type: string
        size:
          type: integer
      required:
        - name
        - objectPath
        - contentType
        - size
...